requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
colorama==0.4.6
pysondb==1.6.7
aioconsole==0.7.0
